+++ src/Pages/Dashboard/AllUser/AllUser.js
@@ -4,7 +4,7 @@ import toast from 'react-hot-toast';
 import Loading from '../../Loading/Loading';
 
 const AllUser = () => {
-    const { data: users = [] ,refetch,isLoading} = useQuery({
+    const { data: users = [], refetch, isLoading } = useQuery({
         queryKey: ['users'],
         queryFn: async () => {
             const res = await fetch('http://localhost:5000/users?userType=Buyer')
@@ -14,24 +14,24 @@ const AllUser = () => {
     })
 
     const handleDeleteUser = (id) => {
+        const proceed = window.confirm('Are you Confirm Delete This User')
+        if (proceed) {
             fetch(`http://localhost:5000/users/${id}`, {
                 method: 'DELETE',
-                headers:{
+                headers: {
                     authorization: `bearer ${localStorage.getItem('accessToken')}`
                 }
-
             })
                 .then(res => res.json())
                 .then(data => {
-                    if(data.deletedCount >0){
-                        toast.success("Delete successfully")  
-                      }
-                    
+                    if (data.deletedCount > 0) {
+                        toast.success("Delete successfully")
+                    }
                     refetch()
                 })
-       
+        }
     }
-    if(isLoading){
+    if (isLoading) {
         return <Loading></Loading>
     }
     return (
